# openapi.yaml
openapi: 3.1.0
info:
  title: SupKube Backend API
  version: "0.9.1.10-alpha"
  description: |
    SupKube 后端 REST API 的机器可读契约。

    **本文件状态**：cross-cutting 契约（认证 / 错误信封 / `/status` / 列表信封 / RBAC）
    已完整定义；resource paths 以**已核对的核心资源**为种子，按模块增量补全。
    **完整端点目录（13 组 ~90 端点）以 [API-REFERENCE.md](API-REFERENCE.md) §7 为准**，
    其权威源是 `supkube-backend/internal/auth/rbac.go` 的 `permissionTable`。

    扩展字段 `x-min-role`：调用该操作所需最低角色（viewer < editor < admin）。
    带 `x-namespace-scoped: true` 的 editor 操作还要过 namespace 作用域检查。
  contact:
    name: SupKube
    url: https://charts.supkube.com
  license:
    name: Proprietary
servers:
  - url: https://{host}/api/v1
    description: 外部（ingress）
    variables:
      host:
        default: supkube.example.com
  - url: http://supkube-backend.{namespace}.svc:8080/api/v1
    description: 集群内
    variables:
      namespace:
        default: supkube
...
security:
  - bearerAuth: []
  - basicAuth: []
paths:
...
  /backups:
    get:
      tags: [backups]
      operationId: listBackups
      summary: 备份列表
      x-min-role: viewer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [backups]
      operationId: createBackup
      summary: 创建备份
      x-min-role: editor
      x-namespace-scoped: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: |
                Backup 规格（透传到 Velero Backup CR 的子集）。
                互斥约束：`snapshotVolumes`(CSI) 与 `defaultVolumesToFsBackup`(filesystem)
                只能选一；`snapshotMoveData` 需 `snapshotVolumes=true`。
              properties:
                name: { type: string }
                includedNamespaces:
                  type: array
                  items: { type: string }
                snapshotVolumes: { type: boolean }
                defaultVolumesToFsBackup: { type: boolean }
                snapshotMoveData: { type: boolean }
      responses:
        "201": { description: Created }
        "400":
          description: 卷模式歧义 / 冲突
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403": { $ref: "#/components/responses/ForbiddenNamespace" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <token>`。token 可为静态 API Token（脚本/MCP/CI）
        或 OIDC JWT（人类经 Dex 登录）。后端按内容自动分派。
    basicAuth:
      type: http
      scheme: basic
      description: htpasswd 兜底（无 IdP 的小集群 / air-gap）。
...
  responses:
    Unauthorized:
      description: 未认证
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "missing or malformed Authorization header" }
...
    ForbiddenNamespace:
      description: 角色不足或 namespace 越权
      content:
        application/json:
          schema: { $ref: "#/components/schemas/RbacError" }
          example:
            error: "namespace not in your scope"
            namespace: "prod"
            scope: ["dev", "staging"]
...
  schemas:
...
    ListResponse:
      type: object
      properties:
        items:
          type: array
          items: { type: object }
        total:
          type: integer
          description: 部分新端点不返回 total，以各端点定义为准
    Error:
      type: object
      description: 传统错误信封（早期端点）
      required: [error]
      properties:
        error: { type: string, description: 给人看的错误文案 }
...
    RbacError:
      type: object
      required: [error]
      properties:
        error: { type: string }
        required: { type: string, enum: [viewer, editor, admin] }
        current: { type: string, enum: [viewer, editor, admin] }
        namespace: { type: string }
        scope:
          type: array
          items: { type: string }
...